✨ Feat|결과창에 get api  추가
