+++ FC.gg/fc_app/src/components/MatchResult.js
@@ -1,22 +1,75 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import "../css/MatchResult.css";
 import loading from "../assets/loading.gif";
 import ball from "../assets/ball.png";
+import axios from "axios";
 
 export default function MatchResult(props) {
-  const { matchdetail, input } = props;
+  const { ouid, input } = props;
   const [expandedId, setExpandedId] = useState(null);
-
-  const toggleExpand = (id) => {
-    setExpandedId((prevId) => (prevId === id ? null : id));
+  const [matchtype, setMatch] = useState(50);
+  const [matchdetail, setMatchdetail] = useState("");
+  const [nickname, setNickname] = useState("");
+  
+  const toggleExpand = (tab) => {
+    setExpandedId((prevId) => (prevId === tab ? null : tab));
   };
 
   // console.log("디리리디리리 : ", Object.entries(matchdetail)[0][1]['my_data'][0]['name']);
   console.log("matchdetail:", matchdetail);
 
+  useEffect(()=>{
+    setNickname(input);
+  }, [input])
+
+  useEffect(()=>{
+    const getMatchdetail = async() => {
+      try {
+        const response1 = await axios.get('https://p0l0evybh6.execute-api.ap-northeast-2.amazonaws.com/dev/Getmatchid', {
+          params: {
+            ouid: ouid,
+            matchtype: matchtype,
+            offset: 0,
+            limit: 10
+          }
+        });
+        setMatch(matchtype);
+  
+        console.log(response1);
+  
+        const matchData = {};
+        console.log('matchid', response1.data);
+  
+        await Promise.all(response1.data.map(async (id) => {
+          console.log('id', id);
+          const response2 = await axios.get('https://p0l0evybh6.execute-api.ap-northeast-2.amazonaws.com/dev/Getmatchdetail', {
+            params: {
+              matchid: String(id),
+              // nickname: sessionStorage.getItem('nickname') 세션 스토리지에 닉네임 넣는 코드가 없어서 그냥 setNickname 있길레 nickname으로 대체했슴다
+              nickname: nickname
+            }
+          });
+          // 응답 데이터를 처리하거나 상태에 저장
+  
+          console.log(response2.data);
+          matchData[id] = response2.data;
+          
+  
+        }));
+      console.log(matchData);
+      setMatchdetail(matchData);
+      console.log(matchdetail);
+      } catch (error) {
+        console.error('Error:', error);
+      }
+    }
+    getMatchdetail();
+  }, [ouid, matchtype, nickname])
+  console.log('MatchResultOuid:', ouid);
+
   return (
     <div className="MatchResultBackground">
-      {matchdetail ? (
+      {matchdetail !== null ? (
         Object.entries(matchdetail).map(([id, data]) => (
           <div className="MatchResultOuterContainer" key={id}>
             <div
@@ -32,7 +85,7 @@ export default function MatchResult(props) {
               onClick={() => toggleExpand(id)}
             >
               <span>
-                {input} {data.my_score} : {data.other_score} {data.other_nick}
+                {nickname} {data.my_score} : {data.other_score} {data.other_nick}
               </span>
               &nbsp; &nbsp; &nbsp;
               <span>{data.my_status}</span>
@@ -41,7 +94,7 @@ export default function MatchResult(props) {
               <div className="MatchResuiltDetailContainer">
                 <div className="MyInformation">
                   <ul>
-                    {data && data.my_data && data.my_player_data ? (
+                    {data && data.my_data && data.my_player_data && (
                       (() => {
                         let combinedData = [];
                         for (let i = 0; i < data.my_data.length; i++) {
@@ -81,14 +134,12 @@ export default function MatchResult(props) {
                           </div>
                         ));
                       })() // 여기서 즉시 실행하지 않음
-                    ) : (
-                      <span>로딩중</span>
                     )}
                   </ul>
                 </div>
                 <div className="OtherInformation">
                   <ul>
-                    {data && data.other_data && data.other_player_data ? (
+                    {data && data.other_data && data.other_player_data && (
                       (() => {
                         let combinedData = [];
                         for (let i = 0; i < data.other_data.length; i++) {
@@ -128,8 +179,6 @@ export default function MatchResult(props) {
                           </div>
                         ));
                       })() // 여기서 즉시 실행하지 않음
-                    ) : (
-                      <span>로딩중</span>
                     )}
                   </ul>
                 </div>
